requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.29
aiosqlite==0.20.0
aiokafka==0.8.0
minio==7.2.3
grpcio==1.62.0
grpcio-tools==1.62.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==6.3.0
sentry-sdk[fastapi]==1.44.0
python-json-logger==2.0.7
kafka-python==2.0.2
alembic==1.13.1